test: Created ApartmentRepositoryTests

+++ hestia-backend/src/main/java/com/jvs0n/hestia/valueobjects/ApartmentHasTenantId.java
@@ -17,11 +17,11 @@ public class ApartmentHasTenantId implements Serializable {
     private static final long serialVersionUID = -4782665201185258150L;
     @NotNull
     @Column(name = "resident_idresident_tenant", nullable = false)
-    private Integer residentIdresidentTenant;
+    private long residentIdresidentTenant;
 
     @NotNull
     @Column(name = "renthistory_idrenthistory", nullable = false)
-    private Integer renthistoryIdrenthistory;
+    private long renthistoryIdrenthistory;
 
 
 }
